feat(Refined Gui first implementation)
Now the Gui system is more ECSish to leverage a more scalable and less chaotic
infrastructure. (Essentially, fonts are now loaded on startup and assigned as
reference inside a component. So now, if an element needs a font to render
some info, the font data is embedded into the entity).